+++ core/runner/src/test/java/com/cognifide/aet/runner/processing/steps/StepManagerTest.java
@@ -24,9 +24,10 @@ import static org.mockito.internal.verification.VerificationModeFactory.times;
 import com.cognifide.aet.communication.api.ProcessingError;
 import com.cognifide.aet.communication.api.metadata.Suite;
 import com.cognifide.aet.communication.api.queues.JmsConnection;
+import com.cognifide.aet.communication.api.wrappers.SuiteRunWrapper;
 import com.cognifide.aet.runner.RunnerConfiguration;
 import com.cognifide.aet.runner.processing.TimeoutWatch;
-import com.cognifide.aet.runner.processing.data.SuiteIndexWrapper;
+import com.cognifide.aet.runner.processing.data.wrappers.RunIndexWrapper;
 import com.cognifide.aet.runner.scheduler.CollectorJobSchedulerService;
 import java.io.Serializable;
 import java.util.Observable;
@@ -67,10 +68,13 @@ public abstract class StepManagerTest {
   protected CollectorJobSchedulerService scheduler;
 
   @Mock
-  protected SuiteIndexWrapper suiteIndexWrapper;
+  protected RunIndexWrapper runIndexWrapper;
 
   @Mock
   protected Suite suite;
+
+  @Mock
+  protected SuiteRunWrapper suiteRunWrapper;
 
   @Mock
   protected MessageConsumer consumer;
@@ -91,9 +95,11 @@ public abstract class StepManagerTest {
     when(connection.getJmsSession()).thenReturn(session);
     when(session.createQueue(anyString())).thenReturn(Mockito.mock(Queue.class));
     when(session.createConsumer(Matchers.<Destination>any(), anyString())).thenReturn(consumer);
-    when(session.createProducer(Matchers.<Destination>any())).thenReturn(sender);
+    when(session.createProducer(
+        Matchers.<Destination>any())).thenReturn(sender);
     when(session.createObjectMessage(Matchers.<Serializable>any())).thenReturn(mockedMessage);
-    when(suiteIndexWrapper.get()).thenReturn(suite);
+    when(runIndexWrapper.get()).thenReturn(suiteRunWrapper);
+    when(suiteRunWrapper.getCorrelationId()).thenReturn(CORRELATION_ID);
     when(suite.getCorrelationId()).thenReturn(CORRELATION_ID);
     when(runnerConfiguration.getMttl()).thenReturn(100L);
     when(runnerConfiguration.getUrlPackageSize()).thenReturn(2);