+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to AET will be documented in this file.
 ## Unreleased
 **List of changes that are finished but not yet released in any final version.**
 
+- [PR-337](https://github.com/Cognifide/aet/pull/337) Added rerun one test or whole suite
+
 ## Version 3.0.0